Add files via upload
Script can be used to add noise to pick files. The generated noise is within the pick uncertainty of each phase. Pick from up to 5 different phases can be manipulated. pickfiles have to be in the popular pasteup format ( Fujie et al., 2008). Pickfiles have to be in the same folder as the script and their names listed in the variable 'pil'. It is similarly crucial to insert the correct uncertainty limits (Nmin, N1max, N2max...).